Create separate test for Redis memory info test, as it is not supported in cluster mode.
