+++ geode-apis-compatible-with-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/string/AbstractAppendIntegrationTest.java
@@ -20,15 +20,12 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import redis.clients.jedis.HostAndPort;
-import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisCluster;
 
 import org.apache.geode.redis.ConcurrentLoopingThreads;
@@ -110,26 +107,6 @@ public abstract class AbstractAppendIntegrationTest implements RedisIntegrationT
   }
 
   @Test
-  public void testAppend_actuallyIncreasesBucketSize() {
-    int listSize = 1000;
-    String key = "key";
-
-    Map<String, String> info = getInfo();
-    Long previousMemValue = Long.valueOf(info.get("used_memory"));
-
-    jedis.set(key, "initial");
-    for (int i = 0; i < listSize; i++) {
-      jedis.append(key, "morestuff");
-    }
-
-    info = getInfo();
-    Long finalMemValue = Long.valueOf(info.get("used_memory"));
-
-
-    assertThat(finalMemValue).isGreaterThan(previousMemValue);
-  }
-
-  @Test
   public void testAppend_withUTF16KeyAndValue() throws IOException {
     String test_utf16_string = "最𐐷𤭢";
     byte[] testBytes = test_utf16_string.getBytes(StandardCharsets.UTF_16);
@@ -155,33 +132,5 @@ public abstract class AbstractAppendIntegrationTest implements RedisIntegrationT
       strings.add(baseString + i);
     }
     return strings;
-  }
-
-  private Map<String, String> getInfo() {
-    Jedis conn = jedis.getConnectionFromSlot(0);
-    try {
-      return getInfo(conn);
-    } finally {
-      conn.close();
-    }
-  }
-
-  /**
-   * Convert the values returned by the INFO command into a basic param:value map.
-   */
-  static Map<String, String> getInfo(Jedis jedis) {
-    Map<String, String> results = new HashMap<>();
-    String rawInfo = jedis.info();
-
-    for (String line : rawInfo.split("\r\n")) {
-      int colonIndex = line.indexOf(":");
-      if (colonIndex > 0) {
-        String key = line.substring(0, colonIndex);
-        String value = line.substring(colonIndex + 1);
-        results.put(key, value);
-      }
-    }
-
-    return results;
   }
 }
